Constraints now a part of performance + some cleanup
Constraint values are now computed in the code, and G is penalized for
violating them.

Qprop now relays performance data back into calc_G so it may be used in
constraint calculation.

Additionally, some of the functions and main were changed to relay more
info to main (G, constraints,flightTime)