Fixed calc errors in runqprop, constraints, main.
Stiffness was being calculated as the inverse of stiffness. Now that
that's changed, the natural frequency constraint can be met.
Additionally, qprop's mass calculation was for the whole craft an not
just a single motor. That was changed to reflect that a single motor
only holds 1/4 of the mass.

Poweruse in calcG was changed to be four times the power use of a single
motor (rather than one).
A residual mass was added in to take into account the rest of the
system.